Consolidate uploads into PDF management
