+++ src/App.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Layout from './components/Layout';
 import HomePage from './components/HomePage';
-import UploadsPage from './components/UploadsPage';
 import DonorsPage from './components/DonorsPage';
 import PdfListPage from './components/PdfListPage';
 import DonationsPage from './components/DonationsPage';
@@ -13,7 +12,6 @@ function App() {
       <Layout>
         <Routes>
           <Route path="/" element={<HomePage />} />
-          <Route path="/uploads" element={<UploadsPage />} />
           <Route path="/pdfs" element={<PdfListPage />} />
           <Route path="/donors" element={<DonorsPage />} />
           <Route path="/donations" element={<DonationsPage />} />
