+++ src/components/Layout.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { Link, useLocation } from 'react-router-dom';
-import { Upload, Home, Users, FileText, HandCoins } from 'lucide-react';
+import { Home, Users, FileText, HandCoins } from 'lucide-react';
 
 // Use an environment variable so the health check works even when the app is served statically
 const API_URL = import.meta.env.VITE_API_URL;
@@ -42,18 +42,6 @@ export default function Layout({ children }: LayoutProps) {
               </Link>
               
               <div className="flex space-x-4 space-x-reverse">
-                <Link
-                  to="/uploads"
-                  className={`flex items-center space-x-2 space-x-reverse px-3 py-2 rounded-md text-sm font-medium transition-colors ${
-                    isActive('/uploads')
-                      ? 'bg-blue-100 text-blue-700'
-                      : 'text-gray-500 hover:text-gray-700 hover:bg-gray-100'
-                  }`}
-                >
-                  <Upload className="h-4 w-4" />
-                  <span>העלאות</span>
-                </Link>
-
                 <Link
                   to="/pdfs"
                   className={`flex items-center space-x-2 space-x-reverse px-3 py-2 rounded-md text-sm font-medium transition-colors ${
